+++ net/minecraft/world/level/BaseSpawner.java
@@ -75,9 +75,9 @@ public abstract class BaseSpawner {
         Level level = this.getLevel();
         BlockPos blockPos = this.getPos();
         if (level.isClientSide) {
-            double d = (float)blockPos.getX() + level.random.nextFloat();
-            double e = (float)blockPos.getY() + level.random.nextFloat();
-            double f = (float)blockPos.getZ() + level.random.nextFloat();
+            double d = (double)blockPos.getX() + (double)level.random.nextFloat();
+            double e = (double)blockPos.getY() + (double)level.random.nextFloat();
+            double f = (double)blockPos.getZ() + (double)level.random.nextFloat();
             level.addParticle(ParticleTypes.SMOKE, d, e, f, 0.0, 0.0, 0.0);
             level.addParticle(ParticleTypes.FLAME, d, e, f, 0.0, 0.0, 0.0);
             if (this.spawnDelay > 0) {
