+++ net/minecraft/world/level/block/BrewingStandBlock.java
@@ -83,9 +83,9 @@ extends BaseEntityBlock {
     @Override
     @Environment(value=EnvType.CLIENT)
     public void animateTick(BlockState blockState, Level level, BlockPos blockPos, Random random) {
-        double d = (float)blockPos.getX() + 0.4f + random.nextFloat() * 0.2f;
-        double e = (float)blockPos.getY() + 0.7f + random.nextFloat() * 0.3f;
-        double f = (float)blockPos.getZ() + 0.4f + random.nextFloat() * 0.2f;
+        double d = (double)blockPos.getX() + 0.4 + (double)random.nextFloat() * 0.2;
+        double e = (double)blockPos.getY() + 0.7 + (double)random.nextFloat() * 0.3;
+        double f = (double)blockPos.getZ() + 0.4 + (double)random.nextFloat() * 0.2;
         level.addParticle(ParticleTypes.SMOKE, d, e, f, 0.0, 0.0, 0.0);
     }
