+++ net/minecraft/world/level/block/EndPortalBlock.java
@@ -52,9 +52,9 @@ extends BaseEntityBlock {
     @Override
     @Environment(value=EnvType.CLIENT)
     public void animateTick(BlockState blockState, Level level, BlockPos blockPos, Random random) {
-        double d = (float)blockPos.getX() + random.nextFloat();
-        double e = (float)blockPos.getY() + 0.8f;
-        double f = (float)blockPos.getZ() + random.nextFloat();
+        double d = (double)blockPos.getX() + (double)random.nextFloat();
+        double e = (double)blockPos.getY() + 0.8;
+        double f = (double)blockPos.getZ() + (double)random.nextFloat();
         double g = 0.0;
         double h = 0.0;
         double i = 0.0;
